pyproject.toml
[tool.poetry]
name = "rclip"
version = "2.1.6"
description = "AI-Powered Command-Line Photo Search Tool"
authors = ["Yurij Mikhalevich <yurij@mikhalevi.ch>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yurijmikhalevich/rclip"
keywords = ["image search", "computer vision", "photography"]
classifiers = [
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Science/Research",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Scientific/Engineering :: Image Recognition",
  "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.11 <3.14"
coremltools = {version = "^9.0", markers = "sys_platform == 'darwin'"}
ftfy = "^6.3.1"
huggingface-hub = "^1.9.0"
numpy = "~2.1.3"
onnxruntime = "^1.24.4"
pillow = "^12.0.0"
pillow-heif = "^1.1.1"
regex = "~=2026.4.4"
requests = "~=2.32"
tqdm = "^4.65.0"
rawpy = "^0.26.1"

[tool.poetry.group.dev.dependencies]
coremltools = {version = "^9.0", markers = "sys_platform != 'windows'"}
open_clip_torch = "^3.2.0"
pyright = {extras = ["nodejs"], version = "^1.1.394"}
pytest = ">=7.2.1,<8.0"
ruff = "^0.9.7"
torch = "2.10.0"
torchvision = "0.25.0"
jinja2 = "^3.1.5"

[tool.poetry.scripts]
rclip = "rclip.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.ruff]
line-length = 120
indent-width = 2
exclude = ["AppDir", "appimage-builder-cache", "benchmarks/datasets"]
